Proof of concept: working with regex!
Took some head scratches and a ton of back and forth changes, but here it is.

Masks are based off regex capture groups. When text matches, it goes into an
array that is then iterated over and inserted into a blank mask to build
the output.

There's still some tweaks that need to be made around cursor positioning.